Refactor mediaElement a bit.
