+++ assets/src/edit-story/elements/text/display.js
@@ -46,10 +46,8 @@ const HighlightWrapperElement = styled.div`
   ${elementFillContent}
   ${elementWithFont}
   ${elementWithTextParagraphStyle}
-  line-height: ${({
-    lineHeight,
-    verticalPadding,
-  }) => getHighlightLineheight(lineHeight, verticalPadding)};
+  line-height: ${({ lineHeight, verticalPadding }) =>
+    getHighlightLineheight(lineHeight, verticalPadding)};
   margin: 0;
   padding: 0;
 `;
